requirements.txt
simple-term-menu>=1.6.1
GitPython>=3.1.40
